feat: add all missing APIs
