+++ src/model.ts
@@ -12,25 +12,90 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { User } from './sdk'
+import {SDK} from './sdk'
+import {AxiosResponse} from "axios";
+import {User} from "./user";
 
-export interface Model {
-  owner: string
-  name: string
-  createdTime: string
-  updatedTime: string
+interface Model {
+    owner: string
+    name: string
+    createdTime: string
+    updatedTime: string
 
-  displayName: string
-  manager: string
-  contactEmail: string
-  type: string
-  parentId: string
-  isTopModel: boolean
-  users?: User[]
+    displayName: string
+    manager: string
+    contactEmail: string
+    type: string
+    parentId: string
+    isTopModel: boolean
+    users?: User[]
 
-  title?: string
-  key?: string
-  children?: Model[]
+    title?: string
+    key?: string
+    children?: Model[]
 
-  isEnabled: boolean
+    isEnabled: boolean
+}
+
+export class ModelSDK extends SDK {
+    public async getModels() {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-models', {
+            params: {
+                owner: this.config.orgName,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Model[]>>
+    }
+
+    public async getModel(id: string) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-model', {
+            params: {
+                id: `${this.config.orgName}/${id}`,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Model>>
+    }
+
+    public async modifyModel(method: string, model: Model) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        const url = `/${method}`
+        model.owner = this.config.orgName
+        const modelInfo = JSON.stringify(model)
+        return (await this.request.post(
+            url,
+            {modelInfo},
+            {
+                params: {
+                    id: `${model.owner}/${model.name}`,
+                    clientId: this.config.clientId,
+                    clientSecret: this.config.clientSecret,
+                },
+            },
+        )) as unknown as Promise<AxiosResponse<Record<string, unknown>>>
+    }
+
+    public async addModel(model: Model) {
+        return this.modifyModel('add-model', model)
+    }
+
+    public async updateModel(model: Model) {
+        return this.modifyModel('update-model', model)
+    }
+
+    public async deleteModel(model: Model) {
+        return this.modifyModel('delete-model', model)
+    }
 }
