+++ src/payment.ts
@@ -12,43 +12,109 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export interface Payment {
-  owner: string
-  name: string
-  createdTime: string
-  displayName: string
+import {SDK} from "./sdk";
+import {AxiosResponse} from "axios";
 
-  // Payment Provider Info
-  provider: string
-  type: string
+interface Payment {
+    owner: string
+    name: string
+    createdTime: string
+    displayName: string
 
-  // Product Info
-  productName: string
-  productDisplayName: string
-  detail: string
-  tag: string
-  currency: string
-  price: number
-  returnUrl: string
+    // Payment Provider Info
+    provider: string
+    type: string
 
-  // Payer Info
-  user: string
-  personName: string
-  personIdCard: string
-  personEmail: string
-  personPhone: string
+    // Product Info
+    productName: string
+    productDisplayName: string
+    detail: string
+    tag: string
+    currency: string
+    price: number
+    returnUrl: string
 
-  // Invoice Info
-  invoiceType: string
-  invoiceTitle: string
-  invoiceTaxId: string
-  invoiceRemark: string
-  invoiceUrl: string
+    // Payer Info
+    user: string
+    personName: string
+    personIdCard: string
+    personEmail: string
+    personPhone: string
 
-  // Order Info
-  outOrderId: string
-  payUrl: string
+    // Invoice Info
+    invoiceType: string
+    invoiceTitle: string
+    invoiceTaxId: string
+    invoiceRemark: string
+    invoiceUrl: string
 
-  state: string
-  message: string
+    // Order Info
+    outOrderId: string
+    payUrl: string
+
+    state: string
+    message: string
+}
+
+export class PaymentSDK extends SDK {
+    public async getPayments() {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-payments', {
+            params: {
+                owner: this.config.orgName,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Payment[]>>
+    }
+
+    public async getPayment(id: string) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-payment', {
+            params: {
+                id: `${this.config.orgName}/${id}`,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Payment>>
+    }
+
+    public async modifyPayment(method: string, payment: Payment) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        const url = `/${method}`
+        payment.owner = this.config.orgName
+        const paymentInfo = JSON.stringify(payment)
+        return (await this.request.post(
+            url,
+            {paymentInfo},
+            {
+                params: {
+                    id: `${payment.owner}/${payment.name}`,
+                    clientId: this.config.clientId,
+                    clientSecret: this.config.clientSecret,
+                },
+            },
+        )) as unknown as Promise<AxiosResponse<Record<string, unknown>>>
+    }
+
+    public async addPayment(payment: Payment) {
+        return this.modifyPayment('add-payment', payment)
+    }
+
+    public async updatePayment(payment: Payment) {
+        return this.modifyPayment('update-payment', payment)
+    }
+
+    public async deletePayment(payment: Payment) {
+        return this.modifyPayment('delete-payment', payment)
+    }
 }
