+++ src/pricing.ts
@@ -12,21 +12,88 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export interface Pricing {
-  owner: string
-  name: string
-  createdTime: string
-  displayName: string
-  description: string
+import {SDK} from "./sdk";
+import {AxiosResponse} from "axios";
 
-  plans?: string[]
-  isEnabled: boolean
-  trialDuration: number
-  application: string
+interface Pricing {
+    owner: string
+    name: string
+    createdTime: string
+    displayName: string
+    description: string
 
-  submitter?: string
-  approver?: string
-  approveTime?: string
+    plans?: string[]
+    isEnabled: boolean
+    trialDuration: number
+    application: string
 
-  state?: string
+    submitter?: string
+    approver?: string
+    approveTime?: string
+
+    state?: string
+}
+
+export class PricingSDK extends SDK {
+    public async getPricings() {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-pricings', {
+            params: {
+                owner: this.config.orgName,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Pricing[]>>
+    }
+
+    public async getPricing(id: string) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-pricing', {
+            params: {
+                id: `${this.config.orgName}/${id}`,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+                pageSize: 1000,
+            },
+        })) as unknown as Promise<AxiosResponse<Pricing>>
+    }
+
+    public async modifyPricing(method: string, pricing: Pricing) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        const url = `/${method}`
+        pricing.owner = this.config.orgName
+        const pricingInfo = JSON.stringify(pricing)
+        return (await this.request.post(
+            url,
+            {pricingInfo},
+            {
+                params: {
+                    id: `${pricing.owner}/${pricing.name}`,
+                    clientId: this.config.clientId,
+                    clientSecret: this.config.clientSecret,
+                },
+            },
+        )) as unknown as Promise<AxiosResponse<Record<string, unknown>>>
+    }
+
+    public async addPricing(pricing: Pricing) {
+        return this.modifyPricing('add-pricing', pricing)
+    }
+
+    public async updatePricing(pricing: Pricing) {
+        return this.modifyPricing('update-pricing', pricing)
+    }
+
+    public async deletePricing(pricing: Pricing) {
+        return this.modifyPricing('delete-pricing', pricing)
+    }
 }
