+++ src/product.ts
@@ -12,26 +12,93 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { Provider } from './provider'
+import {Provider} from './provider'
+import {SDK} from "./sdk";
+import {AxiosResponse} from "axios";
 
-export interface Product {
-  owner: string
-  name: string
-  createdTime: string
-  displayName: string
+interface Product {
+    owner: string
+    name: string
+    createdTime: string
+    displayName: string
 
-  image: string
-  detail: string
-  description: string
-  tag: string
-  currency: string
-  price: number
-  quantity: number
-  sold: number
-  providers?: string[]
-  returnUrl?: string
+    image: string
+    detail: string
+    description: string
+    tag: string
+    currency: string
+    price: number
+    quantity: number
+    sold: number
+    providers?: string[]
+    returnUrl?: string
 
-  state?: string
+    state?: string
 
-  providerObjs?: Provider[]
+    providerObjs?: Provider[]
+}
+
+export class ProductSDK extends SDK {
+    public async getProducts() {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-products', {
+            params: {
+                owner: this.config.orgName,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+                pageSize: 1000,
+            },
+        })) as unknown as Promise<AxiosResponse<Product[]>>
+    }
+
+    public async getProduct(id: string) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-product', {
+            params: {
+                id: `${this.config.orgName}/${id}`,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+                pageSize: 1000,
+            },
+        })) as unknown as Promise<AxiosResponse<Product>>
+    }
+
+    public async modifyProduct(method: string, product: Product) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        const url = `/${method}`
+        product.owner = this.config.orgName
+        const productInfo = JSON.stringify(product)
+        return (await this.request.post(
+            url,
+            {productInfo},
+            {
+                params: {
+                    id: `${product.owner}/${product.name}`,
+                    clientId: this.config.clientId,
+                    clientSecret: this.config.clientSecret,
+                },
+            },
+        )) as unknown as Promise<AxiosResponse<Record<string, unknown>>>
+    }
+
+    public async addProduct(product: Product) {
+        return this.modifyProduct('add-product', product)
+    }
+
+    public async updateProduct(product: Product) {
+        return this.modifyProduct('update-product', product)
+    }
+
+    public async deleteProduct(product: Product) {
+        return this.modifyProduct('delete-product', product)
+    }
 }
