+++ src/provider.ts
@@ -12,50 +12,116 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import {SDK} from "./sdk";
+import {AxiosResponse} from "axios";
+
 export interface Provider {
-  owner: string
-  name: string
-  createdTime: string
+    owner: string
+    name: string
+    createdTime: string
 
-  displayName: string
-  category: string
-  type: string
-  subType: string
-  method: string
-  clientId: string
-  clientSecret: string
-  clientId2: string
-  clientSecret2: string
-  cert: string
-  customAuthUrl: string
-  customTokenUrl: string
-  customUserInfoUrl: string
-  customLogo: string
-  scopes: string
-  userMapping?: Record<string, string>
+    displayName: string
+    category: string
+    type: string
+    subType: string
+    method: string
+    clientId: string
+    clientSecret: string
+    clientId2: string
+    clientSecret2: string
+    cert: string
+    customAuthUrl: string
+    customTokenUrl: string
+    customUserInfoUrl: string
+    customLogo: string
+    scopes: string
+    userMapping?: Record<string, string>
 
-  host: string
-  port: number
-  disableSsl: boolean // If the provider type is WeChat, DisableSsl means EnableQRCode
-  title: string
-  content: string // If provider type is WeChat, Content means QRCode string by Base64 encoding
-  receiver: string
+    host: string
+    port: number
+    disableSsl: boolean // If the provider type is WeChat, DisableSsl means EnableQRCode
+    title: string
+    content: string // If provider type is WeChat, Content means QRCode string by Base64 encoding
+    receiver: string
 
-  regionId: string
-  signName: string
-  templateCode: string
-  appId: string
+    regionId: string
+    signName: string
+    templateCode: string
+    appId: string
 
-  endpoint: string
-  intranetEndpoint: string
-  domain: string
-  bucket: string
-  pathPrefix: string
+    endpoint: string
+    intranetEndpoint: string
+    domain: string
+    bucket: string
+    pathPrefix: string
 
-  metadata?: string
-  idP?: string
-  issuerUrl?: string
-  enableSignAuthnRequest?: boolean
+    metadata?: string
+    idP?: string
+    issuerUrl?: string
+    enableSignAuthnRequest?: boolean
 
-  providerUrl?: string
+    providerUrl?: string
+}
+
+export class ProviderSDK extends SDK {
+    public async getProviders() {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-providers', {
+            params: {
+                owner: this.config.orgName,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Provider[]>>
+    }
+
+    public async getProvider(id: string) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-provider', {
+            params: {
+                id: `${this.config.orgName}/${id}`,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Provider>>
+    }
+
+    public async modifyProvider(method: string, provider: Provider) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        const url = `/${method}`
+        provider.owner = this.config.orgName
+        const providerInfo = JSON.stringify(provider)
+        return (await this.request.post(
+            url,
+            {providerInfo},
+            {
+                params: {
+                    id: `${provider.owner}/${provider.name}`,
+                    clientId: this.config.clientId,
+                    clientSecret: this.config.clientSecret,
+                },
+            },
+        )) as unknown as Promise<AxiosResponse<Record<string, unknown>>>
+    }
+
+    public async addProvider(provider: Provider) {
+        return this.modifyProvider('add-provider', provider)
+    }
+
+    public async updateProvider(provider: Provider) {
+        return this.modifyProvider('update-provider', provider)
+    }
+
+    public async deleteProvider(provider: Provider) {
+        return this.modifyProvider('delete-provider', provider)
+    }
 }
