+++ src/role.ts
@@ -12,15 +12,83 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-export interface Role {
-  owner: string
-  name: string
-  createdTime: string
-  displayName: string
-  description: string
+import {SDK} from './sdk'
+import {AxiosResponse} from 'axios'
 
-  users?: string[]
-  roles?: string[]
-  domains?: string[]
-  isEnabled: boolean
+interface Role {
+    owner: string
+    name: string
+    createdTime: string
+    displayName: string
+    description: string
+
+    users?: string[]
+    roles?: string[]
+    domains?: string[]
+    isEnabled: boolean
 }
+
+export class RoleSDK extends SDK {
+    public async getRoles() {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-roles', {
+            params: {
+                owner: this.config.orgName,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Role[]>>
+    }
+
+    public async getRole(id: string) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        return (await this.request.get('/get-role', {
+            params: {
+                id: `${this.config.orgName}/${id}`,
+                clientId: this.config.clientId,
+                clientSecret: this.config.clientSecret,
+            },
+        })) as unknown as Promise<AxiosResponse<Role>>
+    }
+
+    public async modifyRole(method: string, role: Role) {
+        if (!this.request) {
+            throw new Error('request init failed')
+        }
+
+        const url = `/${method}`
+        role.owner = this.config.orgName
+        const roleInfo = JSON.stringify(role)
+        return (await this.request.post(
+            url,
+            {roleInfo},
+            {
+                params: {
+                    id: `${role.owner}/${role.name}`,
+                    clientId: this.config.clientId,
+                    clientSecret: this.config.clientSecret,
+                },
+            },
+        )) as unknown as Promise<AxiosResponse<Record<string, unknown>>>
+    }
+
+    public async addRole(role: Role) {
+        return this.modifyRole('add-role', role)
+    }
+
+    public async updateRole(role: Role) {
+        return this.modifyRole('update-role', role)
+    }
+
+    public async deleteRole(role: Role) {
+        return this.modifyRole('delete-role', role)
+    }
+}
+
+
